requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-wtf==1.1.1
python-dotenv==1.0.0
pytz==2023.3
werkzeug==2.3.7
email-validator==2.0.0